+++ tests/api/query.py
@@ -3,7 +3,7 @@ import os
 import unittest
 import types
 from time import gmtime, strftime
-from devo.api import Client, Options
+from devo.api import Client, ClientConfig
 
 
 class TestApi(unittest.TestCase):
@@ -11,7 +11,7 @@ class TestApi(unittest.TestCase):
         self.query = 'from demo.ecommerce.data select * limit 1'
         self.app_name = "testing-app_name"
         self.uri = os.getenv('DEVO_API_URL',
-                             'https://apiv2-us.logtrust.com/search/query')
+                             'https://apiv2-us.devo.com/search/query')
         self.key = os.getenv('DEVO_API_KEY', None)
         self.secret = os.getenv('DEVO_API_SECRET', None)
         self.token = os.getenv('DEVO_API_TOKEN', None)
@@ -20,7 +20,7 @@ class TestApi(unittest.TestCase):
         self.comment = os.getenv('DEVO_API_COMMENT', None)
 
     def test_from_dict(self):
-        api = Client.from_dict(
+        api = Client(config=
             {'key': self.key, 'secret': self.secret, 'address': self.uri,
              'user': self.user, 'app_name': self.app_name}
             )
@@ -28,11 +28,11 @@ class TestApi(unittest.TestCase):
         self.assertTrue(isinstance(api, Client))
 
     def test_query(self):
-        options = Options(stream=False, response="json")
+        config = ClientConfig(stream=False, response="json")
 
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=options)
+                     config=config)
 
         result = api.query(query=self.query)
         self.assertIsNotNone(result)
@@ -41,7 +41,7 @@ class TestApi(unittest.TestCase):
     def test_token(self):
         api = Client(auth={"token": self.token},
                      address=self.uri,
-                     options=Options(stream=False, response="json"))
+                     config=ClientConfig(stream=False, response="json"))
         result = api.query(query=self.query)
         self.assertIsNotNone(result)
         self.assertTrue(len(json.loads(result)['object']) > 0)
@@ -49,7 +49,7 @@ class TestApi(unittest.TestCase):
     def test_query_id(self):
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=Options(stream=False, response="json"))
+                     config=ClientConfig(stream=False, response="json"))
         result = api.query(query_id=self.query_id)
         self.assertIsNotNone(result)
         self.assertNotEqual(result, {})
@@ -58,7 +58,7 @@ class TestApi(unittest.TestCase):
     def test_query_yesterday_to_today(self):
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=Options(stream=False, response="json"))
+                     config=ClientConfig(stream=False, response="json"))
         result = api.query(query=self.query,
                            dates={'from': 'yesterday()', 'to': 'today()'})
         self.assertIsNotNone(result)
@@ -67,7 +67,7 @@ class TestApi(unittest.TestCase):
     def test_query_from_seven_days(self):
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=Options(stream=False, response="json"))
+                     config=ClientConfig(stream=False, response="json"))
         result = api.query(query=self.query,
                            dates={'from': 'now()-7*day()', 'to': 'now()'})
         self.assertIsNotNone(result)
@@ -76,7 +76,7 @@ class TestApi(unittest.TestCase):
     def test_query_from_fixed_dates(self):
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=Options(stream=False, response="json"))
+                     config=ClientConfig(stream=False, response="json"))
         result = api.query(query=self.query,
                            dates={'from': strftime("%Y-%m-%d", gmtime()),
                                   'to': strftime(
@@ -88,7 +88,7 @@ class TestApi(unittest.TestCase):
     def test_stream_query(self):
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=Options(response="json/simple"))
+                     config=ClientConfig(response="json/simple"))
         result = api.query(query=self.query)
         self.assertTrue(isinstance(result, types.GeneratorType))
         result = list(result)
@@ -98,10 +98,10 @@ class TestApi(unittest.TestCase):
         """Test the api when the pragma comment.free is used"""
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=Options(response="json",
-                                     stream=False))
-        api.set_user(user=self.user)
-        api.set_app_name(app_name=self.app_name)
+                     config=ClientConfig(response="json",
+                                         stream=False))
+        api.config.set_user(user=self.user)
+        api.config.set_app_name(app_name=self.app_name)
         result = api.query(query=self.query, comment=self.comment)
         self.assertIsNotNone(result)
         self.assertEqual(len(json.loads(result)['object']), 1)
@@ -110,10 +110,10 @@ class TestApi(unittest.TestCase):
         """Test the api when the pragma comment.free is not used"""
         api = Client(auth={"key": self.key, "secret": self.secret},
                      address=self.uri,
-                     options=Options(response="json",
-                                     stream=False))
-        api.set_user(user=self.user)
-        api.set_app_name(app_name=self.app_name)
+                     config=ClientConfig(response="json",
+                                         stream=False))
+        api.config.set_user(user=self.user)
+        api.config.set_app_name(app_name=self.app_name)
         result = api.query(query=self.query)
         self.assertIsNotNone(result)
         self.assertEqual(len(json.loads(result)['object']), 1)